pyproject.toml
[tool.poetry]
name = "Botv1"
version = "0.1.0"
description = ""
authors = ["0xAnkhz <ankhzgamer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.5.2"
requests = "^2.32.3"
python-dotenv = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
